requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
netatmo==1.0.1
python-dotenv==0.13.0
requests==2.23.0
urllib3==1.25.9
